Implement v2 of the Guardrails AI Library with a complete refactor to an OOP design, introducing new provider classes, a factory for provider creation, and improved error handling. This version enhances maintainability and extensibility while ensuring backward compatibility with v1.
