+++ example.js
@@ -1,138 +1,218 @@
-const AI = require('./index.js');
-const dotenv = require('dotenv');
+/**
+ * Example usage of Guardrails AI Library v2
+ *
+ * This demonstrates the OOP-refactored version with proper design patterns.
+ */
 
-// Load environment variables
-dotenv.config();
+const AI = require('./index');
 
-// Example 1: Basic OpenAI usage with proxy
-async function openaiExample() {
-  const ai = new AI({
-    provider: 'openai',
-    modelName: 'gpt-4',
-    system: 'You are a helpful AI assistant.',
-    apiKey: process.env.OPENAI_API_KEY,
-    proxyUrl: process.env.HTTPS_PROXY // Optional proxy configuration
-  });
+/**
+ * Example 1: Using OpenAI
+ */
+async function exampleOpenAI() {
+    console.log('\n=== OpenAI Example ===');
 
-  try {
-    const response = await ai.generate('Explain quantum computing in simple terms');
-    console.log('\nOpenAI Response:', response);
-  } catch (error) {
-    console.error('OpenAI Error:', error.message);
-  }
-}
-
-// Example 2: Anthropic with proxy
-async function anthropicExample() {
-  const claude = new AI({
-    provider: 'anthropic',
-    modelName: 'claude-3-opus-20240229',
-    system: 'You are a helpful AI assistant.',
-    proxyUrl: 'http://your-proxy-server:port' // Direct proxy configuration
-  });
-
-  try {
-    const response = await claude.generate('Write a short poem about artificial intelligence');
-    console.log('\nAnthropic Response:', response);
-  } catch (error) {
-    console.error('Anthropic Error:', error.message);
-  }
-}
-
-// Example 3: Lambda-like usage with proxy from environment
-async function lambdaExample() {
-  const ai = new AI({
-    provider: 'openai',
-    modelName: 'gpt-4',
-    // Proxy will be read from HTTPS_PROXY or HTTP_PROXY environment variable
-  });
-
-  // Simulate Lambda event
-  const event = {
-    body: JSON.stringify({
-      prompt: 'What are the three laws of robotics?',
-      useCase: 'simple'
-    })
-  };
-
-  try {
-    const { prompt } = JSON.parse(event.body);
-    const response = await ai.generate(prompt);
-
-    // Simulate Lambda response
-    console.log('\nLambda-like Response:', {
-      statusCode: 200,
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ response })
+    // Constructor: Only connection details + system prompt (AI's identity)
+    const ai = new AI({
+        provider: "openai",
+        apiKey: process.env.OPENAI_API_KEY || "sk-...",
+        model: "gpt-4",
+        system: "You are a helpful geography teacher."  // AI's persona
     });
-  } catch (error) {
-    // Simulate Lambda error response
-    console.error('\nLambda-like Error Response:', {
-      statusCode: 500,
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        error: error.message,
-        type: error.name
-      })
+
+    // generate(): Request behavior parameters
+    const response = await ai.generate({
+        prompt: "What is the capital of France?",
+        temperature: 0.2,   // Per-request behavior
+        maxTokens: 1000     // Per-request behavior
     });
-  }
+
+    console.log('Provider:', response.provider);
+    console.log('Model:', response.model);
+    console.log('Response:', response.response);
+    console.log('Usage:', response.usage);
 }
 
-// Example 4: Proxy configuration example
-async function proxyExample() {
-  // Example with different proxy configurations
-  const configs = [
-    {
-      name: 'Environment Variable Proxy',
-      config: {
-        provider: 'openai',
-        modelName: 'gpt-4',
-        // Proxy will be read from HTTPS_PROXY environment variable
-      }
-    },
-    {
-      name: 'Direct Proxy URL',
-      config: {
-        provider: 'claude',
-        modelName: 'claude-3-opus-20240229',
-        proxyUrl: 'http://proxy.example.com:8080'
-      }
-    },
-    {
-      name: 'No Proxy',
-      config: {
-        provider: 'openai',
-        modelName: 'gpt-4'
-        // No proxy configuration
-      }
-    }
-  ];
+/**
+ * Example 2: Using Anthropic
+ */
+async function exampleAnthropic() {
+    console.log('\n=== Anthropic Example ===');
 
-  for (const { name, config } of configs) {
+    // Constructor: Connection details only
+    const ai = new AI({
+        provider: "anthropic",
+        apiKey: process.env.ANTHROPIC_API_KEY || "sk-ant-...",
+        model: "claude-3-opus-20240229"
+    });
+
+    // generate(): Behavior parameters
+    const response = await ai.generate({
+        prompt: "Explain quantum computing in simple terms",
+        temperature: 0.2,
+        maxTokens: 1000
+    });
+
+    console.log('Provider:', response.provider);
+    console.log('Response:', response.response);
+}
+
+/**
+ * Example 3: Using AWS Bedrock
+ */
+async function exampleAwsBedrock() {
+    console.log('\n=== AWS Bedrock Example ===');
+
+    // Constructor: Connection details only (provider, apiKey, modelId, region)
+    const ai = new AI({
+        provider: "aws bedrock",
+        apiKey: process.env.AWS_BEDROCK_API_KEY || "ABSK...",
+        modelId: "amazon.nova-lite-v1:0",
+        region: "us-east-1",
+        system: "You are an AWS expert."  // AI's persona
+    });
+
+    // generate(): Behavior parameters
+    const response = await ai.generate({
+        prompt: "What is AWS Bedrock?",
+        temperature: 0.2,
+        maxTokens: 1000
+    });
+
+    console.log('Provider:', response.provider);
+    console.log('Model:', response.model);
+    console.log('Response:', response.response);
+}
+
+/**
+ * Example 4: Using Azure OpenAI
+ */
+async function exampleAzureOpenAI() {
+    console.log('\n=== Azure OpenAI Example ===');
+
+    // Constructor: Connection details (endpoint, deployment, apiVersion are mandatory)
+    const ai = new AI({
+        provider: "azure openai",
+        apiKey: process.env.AZURE_OPENAI_API_KEY || "abc123...",
+        endpoint: "https://my-resource.openai.azure.com/",
+        deployment: "gpt-35-turbo",
+        apiVersion: "2024-04-01-preview"
+    });
+
+    // generate(): Behavior parameters
+    const response = await ai.generate({
+        prompt: "What is Azure OpenAI?",
+        temperature: 1.0,
+        maxTokens: 1000
+    });
+
+    console.log('Provider:', response.provider);
+    console.log('Response:', response.response);
+}
+
+/**
+ * Example 5: Dynamic Provider Selection
+ */
+async function exampleDynamicProvider() {
+    console.log('\n=== Dynamic Provider Example ===');
+
+    // Get supported providers
+    console.log('Supported providers:', AI.getSupportedProviders());
+
+    // Provider configuration from external source (e.g., credentials resolver)
+    // Only connection details in config
+    const providerConfig = {
+        provider: "openai",
+        apiKey: "sk-...",
+        model: "gpt-4",
+        system: "You are a helpful assistant"
+    };
+
+    // Create AI instance dynamically
+    const ai = new AI(providerConfig);
+    console.log('Using provider:', ai.getProviderName());
+
+    // Behavior parameters in generate()
+    const response = await ai.generate({
+        prompt: "Hello, how are you?",
+        temperature: 0.2,
+        maxTokens: 1000
+    });
+    console.log('Response:', response.response);
+}
+
+/**
+ * Example 6: Error Handling
+ */
+async function exampleErrorHandling() {
+    console.log('\n=== Error Handling Example ===');
+
     try {
-      console.log(`\nTesting ${name}:`);
-      const ai = new AI(config);
-      const response = await ai.generate('Hello, how are you?');
-      console.log('Response:', response);
+        // Missing required field (apiKey)
+        const ai = new AI({
+            provider: "openai",
+            modelName: "gpt-4"
+        });
     } catch (error) {
-      console.error('Error:', error.message);
+        console.log('Configuration error caught:', error.message);
     }
-  }
+
+    try {
+        // Unsupported provider
+        const ai = new AI({
+            provider: "unsupported",
+            apiKey: "test",
+            modelName: "test"
+        });
+    } catch (error) {
+        console.log('Provider error caught:', error.message);
+    }
+
+    try {
+        // Missing region for AWS Bedrock
+        const ai = new AI({
+            provider: "aws bedrock",
+            apiKey: "test",
+            modelId: "amazon.nova-lite-v1:0"
+            // Missing: region
+        });
+    } catch (error) {
+        console.log('AWS Bedrock validation error:', error.message);
+    }
 }
 
-// Run all examples
+/**
+ * Run all examples
+ */
 async function runExamples() {
-  console.log('Running AI Examples...\n');
+    console.log('Guardrails AI Library v2 - Examples');
+    console.log('====================================');
 
-  await openaiExample();
-  await anthropicExample();
-  await lambdaExample();
-  await proxyExample();
+    try {
+        await exampleDynamicProvider();
+        await exampleErrorHandling();
+
+        // Uncomment when you have API keys configured:
+        // await exampleOpenAI();
+        // await exampleAnthropic();
+        // await exampleAwsBedrock();
+        // await exampleAzureOpenAI();
+    } catch (error) {
+        console.error('Example error:', error);
+    }
 }
 
-// Run the examples
-runExamples().catch(console.error);
+// Run if executed directly
+if (require.main === module) {
+    runExamples().catch(console.error);
+}
+
+module.exports = {
+    exampleOpenAI,
+    exampleAnthropic,
+    exampleAwsBedrock,
+    exampleAzureOpenAI,
+    exampleDynamicProvider,
+    exampleErrorHandling
+};
+
