+++ providers/BaseProvider.js
